client/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
paho-mqtt>=1.6.1
loguru>=0.7.0
asyncio>=3.4.3 